refactor(cli): group routines into lib/scheduling (Move 3, RUSH-2708)
Move routines.ts (+ test) into new lib/scheduling/ subdir, depth-fix, repoint
57 consumers (imports + string paths, src+tests+bench). Behavior-preserving;
build clean.
