+++ apps/cli/src/lib/__tests__/daemon-self-heal.test.ts
@@ -23,7 +23,7 @@ import {
   getDaemonStatus,
 } from '../daemon.js';
 import { getDaemonDir } from '../state.js';
-import { writeRunMeta, type RunMeta } from '../routines.js';
+import { writeRunMeta, type RunMeta } from '../scheduling/routines.js';
 import { getRunsDir } from '../state.js';
 import { monitorRunningJobs } from '../runner.js';
