+++ apps/cli/src/lib/__tests__/runner.test.ts
@@ -10,9 +10,9 @@ import {
   pinJobBinary,
   resolveRoutineLaunch,
 } from '../runner.js';
-import { readRunMeta } from '../routines.js';
+import { readRunMeta } from '../scheduling/routines.js';
 import { getRunsDir } from '../state.js';
-import type { JobConfig } from '../routines.js';
+import type { JobConfig } from '../scheduling/routines.js';
 import { getBinaryPath, getVersionDir } from '../versions.js';
 import { rotationFailoverChain, type RotateCandidate, type RotateResult } from '../accounting/rotate.js';
 import { detectRateLimit } from '../exec.js';
