+++ apps/cli/src/lib/runner.test.ts
@@ -4,9 +4,9 @@ import * as os from 'os';
 import * as path from 'path';
 import { spawn } from 'child_process';
 import { activeRunSkipStreak, archiveRoutineTranscripts, assertRoutineAccountLocalForPlacement, buildHostDispatchOptions, buildJobCommand, dispatchPlacedJob, executeJob, executeJobDetached, launcherClaimPid, monitorRunningJobs, resolveRoutineLaunch, RoutineAlreadyRunningError, routineSpawnCwd, snapshotRoutineTranscriptBase } from './runner.js';
-import { getRunDir, readRunMeta, writeRunMeta } from './routines.js';
+import { getRunDir, readRunMeta, writeRunMeta } from './scheduling/routines.js';
 import { getVersionHomePath } from './versions.js';
-import type { JobConfig, RunMeta } from './routines.js';
+import type { JobConfig, RunMeta } from './scheduling/routines.js';
 import { hardDeprecationError } from './agents.js';
 import type { RotateCandidate, RotateResult } from './accounting/rotate.js';
 import { saveTask, hostsCacheDir } from './hosts/tasks.js';
