+++ apps/cli/src/lib/triggers/handlers.ts
@@ -15,13 +15,13 @@ import { machineId, normalizeHost } from '../machine-id.js';
 import { safeJoin } from '../paths.js';
 import { pickFleetDevice } from '../routines-placement.js';
 import type { DevicePlatform } from '../devices/registry.js';
-import type { JobConfig, RunMeta, WebhookContext } from '../routines.js';
+import type { JobConfig, RunMeta, WebhookContext } from '../scheduling/routines.js';
 import {
   assertShellSubstitutionSupported,
   readJob,
   substituteWebhookCommand,
   substituteWebhookPrompt,
-} from '../routines.js';
+} from '../scheduling/routines.js';
 import { ensureAgentsDir, getProjectWebhooksDir, getSystemWebhooksDir, getWebhooksDir } from '../state.js';
 import type { AgentId } from '../types.js';
 import type { IncomingWebhook, WebhookSource } from './webhook.js';
