+++ apps/cli/src/lib/triggers/webhook.ts
@@ -23,8 +23,8 @@ import type {
   LinearJobTrigger,
   RunMeta,
   WebhookContext,
-} from '../routines.js';
-import { jobRunsOnThisDevice, listJobs, substituteWebhookPrompt } from '../routines.js';
+} from '../scheduling/routines.js';
+import { jobRunsOnThisDevice, listJobs, substituteWebhookPrompt } from '../scheduling/routines.js';
 import { executeJobDetached } from '../runner.js';
 import { emit } from '../feed/events.js';
 import {
